Revert "move CorsHandler to common"
This reverts commit 1bbc7aa6fe89df015e65d6794d2f1d6c5e8ca7ce.
